courses database

+++ migrations/01_init.sql
@@ -1,15 +1,15 @@
 CREATE TABLE instructors (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     full_name VARCHAR(100) NOT NULL,
     email VARCHAR(100) UNIQUE NOT NULL,
     expertise VARCHAR(100)
 );
 
 CREATE TABLE courses (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     title VARCHAR(150) NOT NULL,
     description TEXT,
-    instructor_id INT NOT NULL,
+    instructor_id BIGINT NOT NULL,
     created_at TIMESTAMP DEFAULT NOW(),
 
     FOREIGN KEY (instructor_id)
@@ -17,16 +17,16 @@ CREATE TABLE courses (
 );
 
 CREATE TABLE students (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     full_name VARCHAR(100) NOT NULL,
     email VARCHAR(100) UNIQUE NOT NULL,
     created_at TIMESTAMP DEFAULT NOW()
 );
 
 CREATE TABLE enrollments (
-    id SERIAL PRIMARY KEY,
-    student_id INT NOT NULL,
-    course_id INT NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    student_id BIGINT NOT NULL,
+    course_id BIGINT NOT NULL,
     enrolled_at TIMESTAMP DEFAULT NOW(),
 
     FOREIGN KEY (student_id)
